requirements.txt
torch==2.3.0
ultralytics ==8.2.16
telebot==0.0.5
easyocr==1.7.1
pillow=10.0.0